services_1/data_generator/requirements.txt
boto3==1.18.0
kafka-python==2.0.2